Error Check, pt 1
This is the SQL that will underlie the error checking function. It
creates a table similar to has_genotype with an additional column
indicating which genotype cannot be rectified with the pedigree. As far
as I can tell, it works, but it will be easier to check once I've come
up with some sort of front-end for displaying the data.